chore: simplify local environment setup
Set up the cockroachdb cluster in single node cluster mode so we only
need one instance of it. This also adds the database initialization as a
docker entrypoint script in `database_entrypoint/a.sql`. This removes
the need to manually connect and initialize the database schema when
brining this up locally. Files in this directory are mounted to the
roach1 db container's docker-entrypoint directory and are executed
automatically in alphabetical order.

Added a .dockerignore file.